+++ NEWS.md
@@ -1,5 +1,27 @@
 # NEWS file for SimDesign
 
+## Changes in SimDesign 2.7
+
+- `SimExtract()` gains a `fuzzy` argument to allow fuzzy matching of error and warning messages.
+  This helps collapse very similar errors messages in the recorded tables,
+  thereby improving how to discern any pattern in the errors/warnings (e.g., Messages such as 
+  *"ERROR: system is computationally singular: reciprocal condition number = 9.63735e-18"* and 
+  *"ERROR: system is computationally singular: reciprocal condition number = 6.74615e-17"* are 
+  effectively the same, and so their number of recorded occurrences should be collapsed)
+
+- Added `AnalyseIf()` function to allow specific analysis function to be included explicitly. Useful
+  when the defined analysis function is not compatible with a row-condition in the `Design` object. 
+  Only relevant when the `analyse` argument was defined as a named list of functions
+
+- The `analyse` argument to `runSimulation()` now accepts a named `list` of functions rather than a 
+  single analysis function. This allows the user to separate the independent analyses into distinct
+  functional blocks rather than having all analyses within the same function, and potentially allows
+  for better modularity. The `debug` argument now also accepts the names of these respective
+  list elements to debug these function definitions quickly
+  
+- `SimFunctions()` gains an `nAnalyses` argument to specify how many analysis functions should 
+  be templated (default is 1, retaining the previous package defaults)
+
 ## Changes in SimDesign 2.6
 
 - Various performance improvements to reduce execution overhead (e.g., `REPLICATION` ID now moved